% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_server.R
\name{dev_server}
\alias{dev_server}
\title{development application server-side}
\usage{
dev_server(input, output, session)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}}
}
\description{
development application server-side
}
