% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compliant_pairs.R
\name{mod_compliant_pairs_server}
\alias{mod_compliant_pairs_server}
\title{compliant_pairs Server Functions}
\usage{
mod_compliant_pairs_server(id, module_data)
}
\arguments{
\item{id}{module id}

\item{module_data}{data from import module}
}
\description{
compliant_pairs Server Functions
}
