% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compliant_pairs.R
\name{mod_compliant_pairs_ui}
\alias{mod_compliant_pairs_ui}
\title{compliant_pairs UI Function}
\usage{
mod_compliant_pairs_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
