% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_import_data.R
\name{mod_import_data_server}
\alias{mod_import_data_server}
\title{import_data Server Functions}
\usage{
mod_import_data_server(id)
}
\arguments{
\item{id}{}
}
\value{
shiny server
}
\description{
import_data Server Functions
}
