% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_import_data.R
\name{mod_import_data_ui}
\alias{mod_import_data_ui}
\title{import_data UI Function}
\usage{
mod_import_data_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
