% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_iso_range.R
\name{mod_iso_range_server}
\alias{mod_iso_range_server}
\title{iso_range Server Functions}
\usage{
mod_iso_range_server(id, module_data)
}
\arguments{
\item{id}{module id}

\item{module_data}{data from import module}
}
\description{
iso_range Server Functions
}
