% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_iso_range.R
\name{mod_iso_range_ui}
\alias{mod_iso_range_ui}
\title{iso_range UI Function}
\usage{
mod_iso_range_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
