% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mard_table.R
\name{mod_mard_table_server}
\alias{mod_mard_table_server}
\title{mod_mard_table Server Functions}
\usage{
mod_mard_table_server(id, module_data)
}
\arguments{
\item{id}{module id}

\item{module_data}{data from import module}
}
\description{
MARD table server function
}
