% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mard_table.R
\name{mod_mard_table_ui}
\alias{mod_mard_table_ui}
\title{mod_table UI Function}
\usage{
mod_mard_table_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
MARD table UI function
}
