% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_modba_graph.R
\name{mod_modba_graph_server}
\alias{mod_modba_graph_server}
\title{modba_graph Server Functions}
\usage{
mod_modba_graph_server(id, module_data)
}
\arguments{
\item{id}{module id}

\item{module_data}{data from import module}
}
\description{
Mod Bland-Altman graph server function
}
