% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pair_type.R
\name{mod_pair_type_server}
\alias{mod_pair_type_server}
\title{mod_pair_type Server Functions}
\usage{
mod_pair_type_server(id, module_data)
}
\arguments{
\item{id}{module id}

\item{module_data}{data from import module}
}
\description{
mod_pair_type Server Functions
}
