% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pair_type.R
\name{mod_pair_type_ui}
\alias{mod_pair_type_ui}
\title{mod_pair_type UI Function}
\usage{
mod_pair_type_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
