% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_risk_category.R
\name{mod_risk_category_server}
\alias{mod_risk_category_server}
\title{risk_level Server Functions}
\usage{
mod_risk_category_server(id, module_data)
}
\arguments{
\item{id}{module id}

\item{module_data}{data from import module}
}
\description{
Risk level table server function
}
