% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_risk_category.R
\name{mod_risk_category_ui}
\alias{mod_risk_category_ui}
\title{risk_level UI Function}
\usage{
mod_risk_category_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
risk level shiny UI module.
}
