% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_risk_grade.R
\name{mod_risk_grade_server}
\alias{mod_risk_grade_server}
\title{risk_grade Server Functions}
\usage{
mod_risk_grade_server(id, module_data)
}
\arguments{
\item{id}{module id}

\item{module_data}{data from import module}
}
\description{
Risk grade table server function
}
