% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_risk_grade.R
\name{mod_risk_grade_ui}
\alias{mod_risk_grade_ui}
\title{risk_grade UI Function}
\usage{
mod_risk_grade_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
Risk grade shiny UI module.
}
