% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_seg_graph.R
\name{mod_seg_graph_server}
\alias{mod_seg_graph_server}
\title{seg_graph Server Functions}
\usage{
mod_seg_graph_server(id, module_data)
}
\arguments{
\item{id}{module id}

\item{module_data}{data from import module}
}
\description{
SEG grid server function
}
