% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_seg_graph.R
\name{mod_seg_graph_ui}
\alias{mod_seg_graph_ui}
\title{seg_graph UI Function}
\usage{
mod_seg_graph_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
